main and sidebar moved to layout folder
